requirements.txt
Flask
google-auth
google-cloud-bigquery
google-cloud-core
jwcrypto
